// src/app/backoffice/models/ShopDTO.ts
export interface ShopDTO {
    id: number;             
    nif: string;            
    isActive: boolean;       
    code: string;           
    name: string;           
    phone: string;          
    email: string;          
    taxes: number;          
    minAmountTaxes: number; 
    address: string;        
    city: string;           
    picture: string;        
    web: string;            
    owner: string;
    ownerId: number;          
    creationDate: string;  
}

 